MNT: update sample file with last_index
